pyproject.toml
[tool.poetry]
name = "autotabloide-ai"
version = "0.1.0"
description = "Offline-first AI-powered tabloid generation system."
authors = ["Antigravity <antigravity@deepmind.google>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.12"
# === Qt Framework (Replaces Flet) ===
pyside6 = "6.6.1"
qasync = "0.27.1"  # Qt-Asyncio integration
# === Database ===
sqlalchemy = "2.0.25"
aiosqlite = "0.19.0"
# === Rendering ===
lxml = "5.1.0"
cairosvg = "2.7.1"
# === AI/ML ===
llama-cpp-python = "0.2.23"
# === Image Processing ===
pillow = "10.2.0"
rembg = "2.0.50"
opencv-python-headless = "4.9.0.80"
onnxruntime = "1.17.0"
# === Utilities ===
pypdf = "4.0.1"
loguru = "0.7.2"
pydantic = "2.6.0"
rapidfuzz = "3.6.0"
psutil = "7.1.3"
chardet = "5.2.0"  # Encoding detection for Excel
openpyxl = "3.1.2"  # Excel parsing
# === Web/Network ===
playwright = "1.41.0"
httpx = "0.26.0"
aiofiles = "23.2.1"
# === Typography ===
fonttools = "4.47.0"
pyphen = "0.15.0"
# === Database Migrations ===
alembic = "1.13.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
# === Optional AI/ML Dependencies ===
# Install with: poetry install -E ai
ai = ["sentence-transformers", "sqlite-vec"]

[tool.poetry.group.ai]
optional = true

[tool.poetry.group.ai.dependencies]
sentence-transformers = "2.2.2"  # RAG embeddings (Vol. IV, Cap. 2.1)
# sqlite-vec = "0.1.0"           # Vector store - install manually

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "black (>=24.0.0)",
    "ruff (>=0.1.0)",
    "mypy (>=1.8.0)"
]

# ==============================================================================
# LINTING (Passo 100 do Checklist)
# ==============================================================================

[tool.black]
line-length = 100
target-version = ['py312']
exclude = '''/(\.eggs|\.git|\.venv|build|dist|alembic|__pycache__)/'''

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "C901"]
exclude = [".eggs", ".git", ".venv", "build", "dist", "alembic"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]
